backend/pyproject.toml
[project]
name = "codecrunchr-backend"
version = "0.0.1"
description = "The backend for the CodeCrunchr app!"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.129.0",
    "gunicorn>=25.0.3",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]
